host QoS: have service curve parameters be integers
It does not make sense to represent fraction of bits, specially when
we get the information from the kernel (though indirectly) and the
kernel external representation is in integer bps.

Reviewed-on: http://gerrit.ovirt.org/34822
Tested-by: Lior Vernia <lvernia@redhat.com>

+++ vdsm/network/tc/_parser.py
@@ -41,28 +41,28 @@ def parse_rate(tokens):
     """Returns a numerical bit representation of the textual rate in tokens"""
     rate = next(tokens)
     if rate[-5:] == 'Gibit':
-        return float(rate[:-5]) * 1024 ** 3
+        return int(float(rate[:-5]) * 1024 ** 3)
     elif rate[-5:] == 'Mibit':
-        return float(rate[:-5]) * 1024 ** 2
+        return int(float(rate[:-5]) * 1024 ** 2)
     elif rate[-5:] == 'Kibit':
-        return float(rate[:-5]) * 1024
+        return int(float(rate[:-5]) * 1024)
     elif rate[-4:] == 'Gbit':
-        return float(rate[:-4]) * 1000 ** 3
+        return int(float(rate[:-4]) * 1000 ** 3)
     elif rate[-4:] == 'Mbit':
-        return float(rate[:-4]) * 1000 ** 2
+        return int(float(rate[:-4]) * 1000 ** 2)
     elif rate[-4:] == 'Kbit':
-        return float(rate[:-4]) * 1000
+        return int(float(rate[:-4]) * 1000)
     else:
-        return float(rate[:-3])
+        return int(float(rate[:-3]))
 
 
 def parse_size(tokens):
     """Returns a numerical byte representation of the textual size in tokens"""
     size = next(tokens)
     if size[-2:] == 'Mb':
-        return float(size[:-2]) * 1024 ** 2
+        return int(float(size[:-2]) * 1024 ** 2)
     elif size[-2:] == 'Kb':
-        return float(size[:-2]) * 1024
+        return int(float(size[:-2]) * 1024)
     else:  # bytes
         return int(size[:-1])
 
@@ -71,11 +71,11 @@ def parse_time(tokens):
     """Returns a numerical µs representation of the textual size in tokens"""
     size = next(tokens)
     if size[-2:] == 'ms':
-        return float(size[:-2]) * 10 ** 3
+        return int(float(size[:-2]) * 10 ** 3)
     elif size[-2:] == 'us':
         return int(size[:-2])
     else:  # s
-        return float(size[:-1]) * 10 ** 6
+        return int(float(size[:-1]) * 10 ** 6)
 
 
 def parse_int(tokens, base=10):
